+++ backend/Makefile
@@ -1,2 +1,3 @@
 run:
+	python manage.py migrate
 	python manage.py runserver
