+++ telegram_bot/handlers/__init__.py
@@ -1,0 +1,1 @@
+from . import admin_handlers
